+++ fund_nav_mcp/error/exceptions.py
@@ -1,0 +1,3 @@
+class DatabaseConnectionError(Exception):
+    """数据库连接错误"""
+    pass